Ngx oauth2 access token/files restructuring (#18)
* provide functions updated

* update importing paths

* remove storage from public api
