+++ projects/mrpachara/ngx-oauth2-access-token/src/lib/provide-access-token.ts
@@ -8,6 +8,12 @@ import {
 } from '@angular/core';
 import { Observable } from 'rxjs';
 
+import { configAccessToken, configRefreshToken } from './functions';
+import {
+  AccessTokenService,
+  Oauth2Client,
+  RefreshTokenService,
+} from './services';
 import { ACCESS_TOKEN_SERVICES, RENEW_ACCESS_TOKEN_SOURCE } from './tokens';
 import {
   AccessTokenConfig,
@@ -16,10 +22,6 @@ import {
   AccessTokenResponseExtractor,
   AccessTokenResponseExtractorInfo,
 } from './types';
-import { RefreshTokenService } from './refresh-token.service';
-import { configAccessToken, configRefreshToken } from './functions';
-import { AccessTokenService } from './access-token.service';
-import { Oauth2Client } from './oauth2.client';
 
 export function provideAccessToken(
   config: AccessTokenConfig,
