+++ projects/mrpachara/ngx-oauth2-access-token/src/lib/provide-jwk.ts
@@ -8,10 +8,10 @@ import {
   makeEnvironmentProviders,
 } from '@angular/core';
 
-import { JwkConfig, JwkFullConfig } from './types';
 import { configJwk } from './functions';
-import { JwkService } from './jwk.service';
+import { JwkService } from './services';
 import { JWK_SERVICES, JWT_INITIALIZED_STATE, JWT_VERIFIERS } from './tokens';
+import { JwkConfig, JwkFullConfig } from './types';
 
 import { JwtEcdsaVerifier } from './jwt-verifiers/jwt-ecdsa.verifier';
 import { JwtHmacVerifier } from './jwt-verifiers/jwt-hmac.verifier';
