+++ projects/mrpachara/ngx-oauth2-access-token/src/lib/types/client.types.ts
@@ -1,4 +1,5 @@
 import { HttpErrorResponse } from '@angular/common/http';
+
 import {
   AccessTokenRequest,
   AuthorizationCodeGrantAccessTokenRequest,
