+++ projects/mrpachara/ngx-oauth2-access-token/src/public-api.ts
@@ -7,20 +7,7 @@ export * from './lib/tokens';
 export * from './lib/errors';
 export * from './lib/functions';
 
-// TODO: may be removed. Storage may use internally.
-export * from './lib/storage';
-
-export * from './lib/oauth2.client';
-export * from './lib/access-token.service';
-export * from './lib/authorization-code.service';
-export * from './lib/refresh-token.service';
-export * from './lib/id-token.service';
-export * from './lib/jwk.service';
-
-// TODO: removed
-// export * from './lib/jwt-verifiers/jwt-hmac.verifier';
-// export * from './lib/jwt-verifiers/jwt-rsassa.verifier';
-// export * from './lib/jwt-verifiers/jwt-ecdsa.verifier';
+export * from './lib/services';
 
 export * from './lib/provide-key-value-pair-storage';
 export * from './lib/provide-oauth2-client';
